test: Significantly improve test suite
This adds a bunch of tests, and refactors most of the current ones to use
the `test_util.setup_pkg` and `test_util.fake_install` functions. This conversion
is being done because it makes the tests significantly easier to reason about
without common function helper abstractions

+++ pkg/lib/util/abstract-bins.sh
@@ -102,7 +102,9 @@ abstract.bins_do_action() {
 			chmod +x "$BPM_INSTALL_BIN/$binName"
 			;;
 		unlink)
-			unlink "$BPM_INSTALL_BIN/$binName"
+			if [ -f "$BPM_INSTALL_BIN/$binName" ]; then
+				unlink "$BPM_INSTALL_BIN/$binName"
+			fi
 			;;
 	esac
 }
